# Add common Contract Benchmarks and Specialize