--- a/gasmodel/Main.hs
+++ b/gasmodel/Main.hs
@@ -7,11 +7,14 @@ import qualified Criterion.Main as C
 
 import Pact.Core.GasModel.InterpreterGas as InterpreterGas
 import Pact.Core.GasModel.BuiltinsGas as BuiltinsGas
+import Pact.Core.GasModel.ContractBench as ContractBench
 
 main :: IO ()
 main = do
+  contractBenches <- ContractBench.allBenchmarks
   C.defaultMain
-    [ InterpreterGas.benchmarks
+    [ contractBenches
+    , InterpreterGas.benchmarks
     , BuiltinsGas.benchmarks
     ]