--- a/gasmodel/Main.hs
+++ b/gasmodel/Main.hs
@@ -9,13 +9,11 @@ import Pact.Core.GasModel.InterpreterGas as InterpreterGas
 import Pact.Core.GasModel.BuiltinsGas as BuiltinsGas
 import Pact.Core.GasModel.ContractBench as ContractBench
 import Pact.Core.GasModel.Serialization as Serialization
-import qualified System.Environment as Env
 
 main :: IO ()
 main = do
-  v <- Env.lookupEnv "RESET_COIN_BENCH_DB"
   C.defaultMain
-    [ ContractBench.allBenchmarks (v == Just "1")
+    [ ContractBench.allBenchmarks
     , InterpreterGas.benchmarks
     , BuiltinsGas.benchmarks
     , Serialization.benchmarks