--- a/gasmodel/Pact/Core/GasModel/BuiltinsGas.hs
+++ b/gasmodel/Pact/Core/GasModel/BuiltinsGas.hs
@@ -10,8 +10,6 @@ import qualified Data.Map.Strict as M
 import qualified Data.Set as S
 import qualified Data.Text as T
 import qualified Data.Vector as V
-import qualified Database.SQLite3 as SQL
-import Control.Lens(over, _2)
 import Control.Monad
 import Data.Bifunctor
 import Data.Default
@@ -902,7 +900,7 @@ benchmarkName = \case
   b -> T.unpack $ coreBuiltinToText b
 
 benchmarks :: C.Benchmark
-benchmarks = C.envWithCleanup mkPactDb cleanupPactDb $ \ ~(pdb, _db) -> do
+benchmarks = C.envWithCleanup mkPactDb cleanupPactDb $ \ ~(pdb, _, _) -> do
   C.bgroup "pact-core-builtin-gas"
     [ C.bgroup (benchmarkName coreBuiltin) benches
     | coreBuiltin <- [minBound .. maxBound]
@@ -910,7 +908,8 @@ benchmarks = C.envWithCleanup mkPactDb cleanupPactDb $ \ ~(pdb, _db) -> do
     , not $ null benches
     ]
   where
-  mkPactDb =
-    over _2 SqliteDbNF <$> unsafeCreateSqlitePactDb serialisePact ":memory:"
+  mkPactDb = do
+    (pdb, db, cache) <- unsafeCreateSqlitePactDb serialisePact ":memory:"
+    pure (pdb, NoNf db, NoNf cache)
 
-  cleanupPactDb (_, SqliteDbNF db) = SQL.close db
+  cleanupPactDb (_, NoNf db, NoNf cache) = unsafeCloseSqlitePactDb db cache