… bench/Bench.hs
… pact-tests/Pact/Core/Test/GasGolden.hs
… pact-tests/Pact/Core/Test/LegacySerialiseTests.hs
… pact-tests/Pact/Core/Test/PersistenceTests.hs
… pact-tests/Pact/Core/Test/ReplTests.hs
… pact-tests/Pact/Core/Test/SizeOfTests.hs
… pact-tests/Pact/Core/Test/StaticErrorTests.hs
… pact-tng.cabal
… pact/Pact/Core/Compile.hs
… pact/Pact/Core/Environment/Types.hs
… pact/Pact/Core/Environment/Utils.hs
… pact/Pact/Core/Errors.hs
… pact/Pact/Core/Evaluate.hs
… pact/Pact/Core/Gas/Types.hs
… pact/Pact/Core/Guards.hs
… pact/Pact/Core/IR/ConstEval.hs
… pact/Pact/Core/IR/Desugar.hs
… pact/Pact/Core/IR/Eval/CEK.hs
… pact/Pact/Core/IR/Eval/CEK/Types.hs
… pact/Pact/Core/IR/Eval/CEK/Utils.hs
… pact/Pact/Core/IR/Eval/CoreBuiltin.hs
… pact/Pact/Core/IR/Eval/Direct/Evaluator.hs
… pact/Pact/Core/IR/Eval/Direct/ReplBuiltin.hs
… pact/Pact/Core/IR/Eval/Direct/Types.hs
… pact/Pact/Core/IR/Eval/Runtime/Types.hs
… pact/Pact/Core/IR/Eval/Runtime/Utils.hs
… pact/Pact/Core/Info.hs
… pact/Pact/Core/Interpreter.hs
… pact/Pact/Core/Legacy/LegacyPactValue.hs
… pact/Pact/Core/Names.hs