This PR adds pact-5 benchmarks for common contracts like coin, and marmalade.

As a result of the benchmarking, this PR monomorphizes types across the repo for a significant speed improvement, as seen in the following:  
Before single monad: 
```
benchmarking PactBenchmarks/CoinTransfer/EvalOnly(transfer) from SenderA to SenderB using: CEK
time                 293.8 μs   (293.5 μs .. 294.0 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 293.7 μs   (293.6 μs .. 293.9 μs)
std dev              544.2 ns   (457.3 ns .. 674.9 ns)

benchmarking PactBenchmarks/CoinTransfer/EvalOnly(transfer) from SenderA to SenderB using: CEKSpec
time                 92.43 μs   (92.32 μs .. 92.52 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 92.30 μs   (92.24 μs .. 92.38 μs)
std dev              218.7 ns   (172.5 ns .. 268.0 ns)
```
After:
```
benchmarking PactBenchmarks/CoinTransfer/EvalOnly(transfer) from SenderA to SenderB using: CEK
time                 56.43 μs   (56.37 μs .. 56.49 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 56.51 μs   (56.45 μs .. 56.57 μs)
std dev              175.0 ns   (142.5 ns .. 244.3 ns)
```

Checklist:
- [x] Coin contract setup
- [x] Coin DB pre-population
- [x] `coin.transfer` benchmarks: CEK and Direct-style
- [x] Monad monomorphization
- [ ]  `coin.transfer-create` benchmarks.

TODO: Other contract benchmarks.